Cleanup: remove old DataList and rename Datalist2 to DataList
